ML/requirements.txt
# HealthPath ML Service - Production Dependencies
# ================================================

# Core ML Libraries
numpy>=2.0.0
pandas>=2.1.0
scikit-learn>=1.5.0
xgboost>=2.0.0
joblib>=1.3.0

# Web Framework
flask>=2.3.0,<4.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Visualization (for training reports)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
